index thread support
attempted to add support for extracting stories from an index thread & naming them based on their captions in that thread, as well as small improvements in other areas.

this *appears* to work but I expect there are situations that will break it. as with the rest of this script, it's a work in progress.